+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-copy-tracker.js
@@ -1,7 +1,3 @@
-window.GOVUK = window.GOVUK || {}
-window.GOVUK.analyticsGa4 = window.GOVUK.analyticsGa4 || {}
-window.GOVUK.analyticsGa4.analyticsModules = window.GOVUK.analyticsGa4.analyticsModules || {};
-
 (function (analyticsModules) {
   'use strict'