+++ app/assets/javascripts/govuk_publishing_components/analytics-ga4/ga4-core.js
@@ -1,6 +1,3 @@
-window.GOVUK = window.GOVUK || {}
-window.GOVUK.analyticsGa4 = window.GOVUK.analyticsGa4 || {};
-
 (function (analytics) {
   'use strict'