+++ app/assets/javascripts/govuk_publishing_components/components/accordion.js
@@ -1,7 +1,5 @@
 // This component relies on JavaScript from GOV.UK Frontend
 // = require govuk/components/accordion/accordion.bundle.js
-window.GOVUK = window.GOVUK || {}
-window.GOVUK.Modules = window.GOVUK.Modules || {}
 window.GOVUK.Modules.GovukAccordion = window.GOVUKFrontend.Accordion;
 
 (function (Modules) {