+++ app/assets/javascripts/govuk_publishing_components/components/password-input.js
@@ -1,5 +1,3 @@
 // This component relies on JavaScript from GOV.UK Frontend
 // = require govuk/components/password-input/password-input.bundle.js
-window.GOVUK = window.GOVUK || {}
-window.GOVUK.Modules = window.GOVUK.Modules || {}
 window.GOVUK.Modules.GovukPasswordInput = window.GOVUKFrontend.PasswordInput