+++ app/assets/javascripts/govuk_publishing_components/components/reorderable-list.js
@@ -1,7 +1,4 @@
 //= require sortablejs/Sortable.js
-window.GOVUK = window.GOVUK || {}
-window.GOVUK.Modules = window.GOVUK.Modules || {};
-
 (function (Modules) {
   function ReorderableList ($module) {
     this.$module = $module